fix(versions): fixes dependencies in the concrete ecosystem.
The development of the different crates of the ecosystem are not meant to be synchronized. This means that we want the whole workspace to compile in every case, even when one crate upper in the dependency graph was not updated to match the api of its dependencies. For that we had to replace dependencies to "=X.X.X" to exactly match versions, and we also had to add a _patch_ section to the workspace cargo.toml. Finally, all the versions of all crates were bumped, and a new version for each crate must be published.