refactor(decomposition): refactors the signed decomposition module.
This commit changes the way signed decomposition is performed in concrete.

A new structure `SignedDecomposer` is introduced, which represent a given decomposition. It allows to compute the closest value representable by the decomposition, for values and tensors. It also allows to access iterators to perform the actual decomposition.

For values, the `SignedDecompositionIter` implements `Iterator` and yield a new term of the decomposition at each iteration. The terms of the decomposition are yielded in the reversed order (from larger to smaller). If this iterator was not iterated on, it can be consumed to compute the recomposition of the initial value using the `recompose` method.

Decomposition terms are represented by `DecompositionTerm` objects, which contains the actual level of the term and its value. It also makes it possible to convert a term to its summand counterpart to compute the recomposition of a value (`to_summand` method).

For tensors, the `TensorSignedDecompositionIter` plays the role of an iterator. At every call to the `next_term` method, a tensor backed by a slice is returned to the user. This tensor contains all the terms of the decomposition for the input and for the current level. This slice points to a vec contained in the `TensorSignedDecompositionIter`, which means that the tensor term must be dropped before the next call to `next_term`. Sadly this can not be done the way `Iterator` is currently umplemented (that would mean GATs). For this reason, the `next_term` method must be explicitly called during the iteration.

Tensor decomposition terms are represented by a `DecompositionTermTensor` structure, which contain the current level of the decomposition, and maks it possible to fill another tensor with every values truned to summands.