refactor(bootstrap): refactors the bootstrapping keys.
This commit contains a refactoring of the bootstrap module of concrete-core. The `cross` module was removed, and the functions which existed there were merged inside the structures themselves.

A new type `StandardBootstrapKey` represents a bootstrap key in the standard domain. This key is not yet able to perform an actual bootstrap though. A new type `FourierBootstrapKey` represents a bootstrap key in the fourier domain for which the bootstrap is performed using the fourier transform. Bootstrapping itself is now performed via a `Bootstrap` trait which expose a `bootstrap` method.

Importantly, all the buffers used by `FourierBootstrapKey` now live inside the bootstrap key itself. This allows to reuse the fft plans and the buffers from bootstrap to bootstrap.

Also, `concrete-core` now depends on the new `concrete-fftw` package. This allowed to simplify the serialization of the different bootstrap keys.

Finally, benchmarks, which were broken, were fixed, and various warnings were fixed throughout the codebase.