refactor(secret_keys): allow different key kinds
This commit makes it possible to use different kinds of secret keys in
concrete. We now have access to the kinds `BinaryKeyKind`,
`TernaryKeyKind`, `GaussianKeyKind`, and `UniformKeyKind`. Those kinds
are represented by empty types all implementing the same marker trait
`KeyKind`.

The `LweSecretKey` and `GlweSecretKey` types now have a generic type
parameter to encode the kind of key contained (which must implement the
`KeyKind` marker trait). Also, the secret keys can now use unsigned
integers of various precisions as data-type, where it used to
only work with `bool`.

A new random generator specialy used for generating secret keys is
introduced, the `SecretRandomGenerator` structure. Both this generator
and the one used to encrypt the ciphertexts (the
`EncryptionRandomGenerator`) are now located in the `crypto::secret`
module.