fix(ci): restore fftw system dependency.
The dependency on the system installation of libfftw was removed from
the CI in commit 6743dce, because the dependency on the system fftw
library was removed from `concrete-core=0.1.10`. It was a mistake, as
the `concrete` crate still depends on `concrete-core=0.1.9`, which
itself depends on `fftw-sys`, with the `system` features, which requires
libfftw to be installed.

This commit restore the installation of the system fftw in the CI, to
solve this issue.