feat(concrete_core): add lwe sample extraction to the public API
This commit brings lwe sample extraction to the public API of
`concrete-core`. The methods
`GlweCiphertext::fill_lwe_with_sample_extraction` and
`LweCiphertext::fill_with_glwe_sample_extraction` make it possible to
extract any coefficient of a `GlweCiphertext`, into a properly sized
`LweCiphertext`. The elements of the mask are rearranged in such a way
that reinterpreting the `GlweSecretKey` into a `LweSecretKey` via the
`GlweSecretKey::into_lwe_secret_key` method (without making any change
in the order of the underlying array of elements), allows to decrypt
the newly created `LweCiphertext`.

This commit also adds three new methods to the `Tensor` type, the
`reverse`, `rotate_left`, and `rotate_right` methods.