fix(ci): fixes broken ci
The CI was broken because of two issues fixed by this commit. The first problem was that some links in the documentation were broken. The second problem was that it was using a version of `concrete-fftw` which was incompatible with code coverage computation.

Also, in order to make it simpler for users to reproduce the commands of the CI, we added a new binary target to the ecosystem, the `concrete-tasks` crate. This one is not supposed to be published, and contains commands which allows to perform various tasks on the repository, such as checking, testing, building, and so on. Moreover, we added a cargo alias, which makes it possible to use `cargo xtask ...` to execute the binary target. The actions files were updated accordingly.

Finally, the clippy lints were modified to fail on any warning. This allows to ensure that the lints must be thoroughly studied, and eventually discarded by an `allow` macro. Consequently, the code was modified to please clippy.