feat(ecosystem): add a library for homomorphic Boolean circuits (zama-ai/concrete_internal#79)
Add a new crate implementing homomorphic logic gates. Algorithms come from TFHE's gate bootstrapping.
There are two different types of keys: `ClientKey` which is a private key and `ServerKey` which is a public key used for computing homomorphic algorithms on a server. The `ClientKey` is used to encrypt and decrypt bits. It is possible to compute AND, OR, NAND, XOR, XNOR, NOR, NOT and MUX gates.

Co-authored-by: ilaria-zama <ilaria.chillotti@zama.ai>
Co-authored-by: J-B Orfila <jb.orfila@zama.ai>
Co-authored-by: aPere3 <alexandre.pere@zama.ai>
Co-authored-by: aquint-zama <alexandre.quint@zama.ai>
Co-authored-by: Samuel Tap <samuel.tap@cryptoexperts.com>