feat(concrete_core): add a modular api on top of concrete-core
A new architecture and api are brought to the `concrete-core` crate.

The api is fully specified in a new module named `specification`, which contains traits for datatypes and operators. It represents in an abstract way, the concrete homomorphic instruction set. This specification is based on two ideas. Objects implementing the `AbstractEntity` trait plus one child `*Entity` trait, represent an homomorphic datatype of a given kind, under given representation. Objects implementing the `AbstractEngine` trait plus any number of `*Engine` traits, can be used to perform operations on entities.

This api can now be implemented a-la-carte by different modules, which can be conditionally included in the build of the user. Those modules are available in the `backends` module, and are activated with `backend_*` feature flags. this commit only brings a single backend, named `core`, which contains nothing more but the previous `concrete-core` code.

Signed-off-by: Alexandre Péré <alexandre.pere@zama.ai>