feat(concrete_core): add engine traits to retrieve arbitrary values.
This commit adds new engine traits which allows to retrieve arbitrary
values from cleartext and plaintext entities (and their vectors alike).
This makes it possible for the user to retrieve the output of
decryptions in a better suited format (usually raw integers).

We added the following engines:
+ `CleartextDiscardingRetrievalEngine`
+ `CleartextRetrievalEngine`
+ `CleartextVectorDiscardingRetrievalEngine`
+ `CleartextVectorRetrievalEngine`
+ `PlaintextDiscardingRetrievalEngine`
+ `PlaintextRetrievalEngine`
+ `PlaintextVectorDiscardingRetrievalEngine`
+ `PlaintextVectorRetrievalEngine`

We also added an implementation of those engine traits to the
`CoreEngine` of the `core` backend.