feat(concrete_correctness): add an application to test core correctness
This commit brings a generic statistical testing application to the ecosystem. This application contains generic testing functions, aimed at verifying that the statistical properties of the operators implemented in `concrete-core` follow our models (embodied by the formulas implemented in `concrete-npe`). The generic nature of the test functions implemented here, makes it possible to test every operators of the `concrete-core` library, using the same functions. Then, testing a new backend mainly consists in appropriately instantiating the tests.