feat(ci): release on tag pushed (linux only)
The release workflow should, whenever a tag is pushed,
create a draft release on github (that should be manually
checked by one of us to make the actual release).
The release should contain wheels for python 3.8, 3.9, and 3.10,
as well as a tarball containing the zamacompiler and required
libraries (and installation guide is provided).
The draft release will also have an auto generated release note
that can be modified later.