chore(core): use rustfmt::skip as inner attribute
Using `#[rustfmt::skip]` on the `pub mod specification`
made rustfmt skip the whole specification module.

What we wanted is just the modules not to be reordered
in `concrete-core/src/lib.rs`, so we instead use
the inner attribute `#![rustfmt::skip]` so that the file is not
formatted, but the modules declared via pub mod are.

We use`cfg_attr` as custom inner attributes are only available on nightly