feat(concrete_core): rework fixtures to increase performance with reuse
The fixture crate re-generated all the inputs for each test that was
performed. For performance reasons though, we needed to be able to
reuse some of those inputs across executions (for example, perform
all the tests with one or several BSK, not regenerating the BSK for
each sample).

This commit introduces a different structure for the fixture, which
allows that.