fix(compiler): Fix calculation of indexes for result memrefs with non-zero offsets
The function `tensorDataFromMemRef` reading elements from a result
memref and copying them to an instance of `TensorData` fails to handle
memrefs with non-zero offsets as it adds the offset twice when
calculating indexes. This may result in out-of-bounds memory accesses
and incorrect results.

This patch fixes the calculation of indexes by removing the second
addition of the offset.