refactor(concrete): merge dynamic and static types
This adds an "Id" system, where each type of parameters has an id tied
to it. The id is stored in the ciphertext and the id contains the logic
of where to look for the keys.

BREAKING_CHANGE: The parameters `DynShortIntParameters` type changed.

A lot of changes made in this refactor are likely to be "technically breaking", in
that, for most user won't break their code, but can still break for other users
doing more advanced stuff.