chore: disable license check in pcc
It needs to be disabled because we use different versions of dependencies depending on the python version.

This system was not ideal anyway, it was just checking for string equality, whereas it should have checked if the licenses are accepted without looking into a pre-committed file.

Another PR will follow to improve the situation around licenses.