fix(concrete_cuda): make sure r > 0 in the wop PBS
Sometimes the optimizer provides inconsistent parameter sets for the wop
PBS depending on how many inputs & bits to extract we're going to need
for a certain precision. This results in r < 0 and until now the
corresponding error in concrete-cuda was very hard to understand. This
commit fixes this behavior.

Also, when introducing the support for k > 1 we forgot to update the
checks on the low lat PBS. This commit also fixes them.