test(compiler): Let entire Makefile rules for tests / benchmarks fail if subcommand fails
The commands for the Makefile rules `run-end-to-end-tests` and
`run-cpu-benchmarks` are generated using the `$(foreach)` builtin
function with a trailing `;` for each subcommand. The result is a
sequence of commands separated by `;`, which only fails if the last
command fails. This causes any code for which the last test /
benchmark passes, but for which any of the preceding subcommands fails
to pass the CI and thus to leave failing tests undiscovered.

This commit adds a trailing `&& exit $$?` to each command, such that
the entire sequence fails if a single command fails.